hotel_room completed

+++ app/hotel/services.py
@@ -7,30 +7,40 @@ from .models import HotelRoom
 from .repositories import HotelRoomRepository
 from .serializers import HotelRoomSerializer
 
+
 class HotelRoomService:
     """
     Сервисный слой для работы с номерами отеля
     """
+
     def __init__(self):
         self.repository = HotelRoomRepository()
 
+    def room_exists(self, room_id: int) -> Response | HotelRoom:
+        try:
+            return self.repository.get_one(room_id)
+        except HotelRoom.DoesNotExist:
+            return Response({
+                'status': 'error',
+                'message': 'Комната не найдена'
+            }, status=status.HTTP_404_NOT_FOUND)
+
     def get_room(self, room_id: int) -> Response:
         """Получить номер по id"""
-        try:
-            result = self.repository.get_one(room_id)  # Вызывает .get()
-            serializer = HotelRoomSerializer(result)
-            return Response({'data': serializer.data}, status=status.HTTP_200_OK)
-        except HotelRoom.DoesNotExist:
-            return Response({'error': 'Комната не найдена'}, status=status.HTTP_404_NOT_FOUND)
+        room = self.room_exists(room_id)
+        if type(room) == Response:
+            return room
+        serializer = HotelRoomSerializer(room)
+        return Response({'data': serializer.data}, status=status.HTTP_200_OK)
 
-    def get_room_list(self, sort_by:  str = 'created_at', order: str = 'asc') -> Response:
+    def get_room_list(self, sort_by: str = 'created_at', order: str = 'asc') -> Response:
         """Получить список номеров"""
         rooms = self.repository.get_all(sort_by, order)
+
         serializer = HotelRoomSerializer(rooms, many=True)
         return Response({
             'data': serializer.data
         })
-
 
     def create_room(self, data: Dict[str, Any]) -> Response:
         """Создать новый номер отеля"""
@@ -52,13 +62,12 @@ class HotelRoomService:
 
     def delete_room(self, room_id: int) -> Response:
         """Удалить номер отеля"""
-        try:
-            room = self.repository.get_one(room_id)
-        except HotelRoom.DoesNotExist:
-            return Response({'error': 'Номер не найден'}, status=status.HTTP_404_NOT_FOUND)
-
+        room = self.room_exists(room_id)
+        if type(room) is Response:
+            return room
         room.delete()
         return Response({'message': f'Номер {room_id} успешно удален'}, status=status.HTTP_204_NO_CONTENT)
 
 
-
+class HotelBillingService:
+    pass
